+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Prevent context processor error when WEB_ANALYTICS_ID setting is not defined.
 - Fix autocomplete views tests that may failed because of result order.
 - Fix section tile title with richtext and special character escaping.
+- Change blogpost detail template to display author even in published mode
+  except if its placeholder is empty.
 
 ## [2.13.0] - 2022-02-18